refactor: create `foundation-args` library
We're already doing argument parsing in 2 different places to get the
location of configuration files, so let's update this to be done in a
single place instead.

This change:
* Adds a `foundation-args` library
* Updates everything to use it
