+++ apps/foundation/args/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "foundation-args"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+color-eyre = "0.6.5"
+pico-args = "0.5.0"
